Refactor course creation components and implement new lesson builder
- Removed the old BlockToolbar component and replaced it with a new implementation.
- Introduced ContentCreateDropdown for creating lessons and quizzes.
- Developed CourseBuilder for managing course details and lesson inputs.
- Added AddLessonBlock for handling lesson sections and content creation.
- Created LessonBuilder for editing lesson content with an integrated editor.
- Implemented hooks for managing lesson state and interactions.
- Added various text block components for structured content creation.
- Enhanced folder management with a dropdown menu for actions.
- Updated the overall structure and organization of course creation components.
