+++ src/app/courses/create/lesson-builder/page.tsx
@@ -1,4 +1,4 @@
-import LessonBuilder from '@/features/courses/_components/course_builder/create-content'
+import LessonBuilder from '@/features/courses/components/course_builder/create-content'
 import React from 'react'
 
 export default function page() {
