+++ src/app/courses/create/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import CoursesBuilder from '@/features/courses/_components/course_builder/course-builder'
+import CoursesBuilder from '@/features/courses/components/course_builder/course-builder'
 
 export default function CoursesPage() {
   return <CoursesBuilder />
