+++ src/app/courses/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import SharedLayout from '@/components/layout/shared/layout'
-import CoursesBoard from '@/features/courses/_components/course_creation/course'
+import CoursesBoard from '@/features/courses/components/course_creation/course'
 
 export default function CoursesPage() {
   return (
