+++ src/features/courses/components/course_builder/course-builder.tsx
