+++ src/features/courses/components/course_builder/create-content.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Editor from '@/features/courses/_components/course_creation/editor/editor'
+import Editor from '@/features/courses/components/course_creation/editor/editor'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { IoArrowBack } from 'react-icons/io5'
