+++ src/features/courses/components/course_creation/editor/editor.tsx
@@ -2,10 +2,10 @@
 
 import React, { useState } from 'react'
 import { v4 as uuid } from 'uuid'
-import BlockToolbar from '@/features/courses/_components/course_creation/tool-bar'
-import AddTextBlockModal from '@/features/courses/_components/course_creation/add-text-block-modal'
-import ParagraphBlock from '@/features/courses/_components/course_creation/blocks/text_blocks/paragraph-block'
-import HeadingWithParagraphBlock from '@/features/courses/_components/course_creation/blocks/text_blocks/paragraph-with-headting'
+import BlockToolbar from '@/features/courses/components/course_creation/tool-bar'
+import AddTextBlockModal from '@/features/courses/components/course_creation/add-text-block-modal'
+import ParagraphBlock from '@/features/courses/components/course_creation/blocks/text_blocks/paragraph-block'
+import HeadingWithParagraphBlock from '@/features/courses/components/course_creation/blocks/text_blocks/paragraph-with-headting'
 import { TextBlock } from '@/types'
 import { TEXT_BLOCKS } from '../blocks/text_blocks/text-block-data'
