+++ src/features/courses/components/course_creation/tool-bar.tsx
